requirements.txt
asgiref==3.7.2
certifi==2023.7.22
distlib==0.3.8
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
mysqlclient==2.2.1
platformdirs==4.1.0
protobuf==4.24.4
PyAudio==0.2.14
pydub==0.25.1
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
rptree==0.1.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
wheel==0.41.3